Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (4 comments)
> 
> For my own understanding, what is the meaning of uRPF when applied to tx? 
> 
> Is it sufficient to clone the ip4 tests as a template for ip6?
> Is there a need for tests for:
> - interfaces with secondary addresses?
> - dhcp proxy on uRPF interface?

on tx it means "do not allow packets to leave my network that are not sourced by a prefix that belongs to my network". we enforce this by saying, do not forward a packet out of an interface whose source address is routable thru that interface - i.e. don't send packets back where they would come from.

i will clone the ip4 tests for ip6.

no need for tests with odd interface addresses since the DP checks are against the FIB not against the interface-IP address DB.

Patch-set: 2