Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> what's the motivation?

Static analysis from SonarCloud flags this as an issue which I agree with. Unlike most of the pep8 type things we enforce, this provides a degree of value.  It is the equivalent discipline found within the c code when 
#define foreach_gbp_vxlan_input_next         \
  _(DROP, "error-drop")                      \
  _(L2_INPUT, "l2-input")                    \
  _(IP4_INPUT, "ip4-input")                  \
  _(IP6_INPUT, "ip6-input")
is done.

Other sites supporting the value of this practice can be found at: 
https://stackoverflow.com/questions/10662299/should-i-avoid-magic-strings-as-possible
https://code.tutsplus.com/tutorials/refactoring-legacy-code-part-2-magic-strings-constants--cms-20527
https://www.jetbrains.com/help/pycharm/extract-constant.html
https://sourcery.ai/blog/five-refactoring-tips/

It is a byproduct of our practice of copying a test for reuse instead of refactoring the original.  I personally don't have time to fully refactor the code base.  Contributions are welcome!

I have another wave planned where test specific constants will be refactored out to the top of the individual unit test files.

Patch-set: 5