misc: require indent 2.2.11 for make checkstyle, allow local temporary install
Some platforms (e.g. Ubuntu 20.08) have GNU indent 2.2.12, which fixes a few
bugs and adds a few new bugs and differences in behavior.
That causes the inconsistency in make checkstyle/make fixstyle behavior,
vs. the style used in the codebase, potentially requiring a large
change just to accomodate a new behavior for the indent.

This patch adds a check that the indent version is 2.2.11,
and adds option to easily install a local copy that will be used
just for the checkstyle/fixstyle.

Type: feature
Change-Id: Icdb3c821d82c1d59090c8d86287a0b5ac295435e
Signed-off-by: Andrew Yourtchenko <ayourtch@gmail.com>