build: option to also build vppctl_static - a statically linked vppctl
For containerized setups, it is nice to not have to install the
libvppinfra dependency - with statically linked executable,
one can just grab the vppctl and start talking to the vpp
via an exposed socket.

This commit does not change the behavior of the default build.
To build the vppctl_static, one needs to have variable
BUILD_STATIC_VPPCTL=1 in the build environment.

That causes a static version of libvppinfra to be built
along with usual artifacts, and then vppctl_static is created.
Neither of the two are installed into packages.

Type: make

Change-Id: If52f91cc28ff15f121a2fc2e138ba93feff95a73
Signed-off-by: Andrew Yourtchenko <ayourtch@gmail.com>