Update patch set 1
Patch Set 1: Code-Review+1

> vpp-csit-verify-perf-1908-3n-hsw/6/ : FAILURE (skipped)

Finally, after some cherry-picking on CSIT side, this run proves this is a fix.
The "current" build (with this Change) test passes [0], but the "parent" build (without this) still fails [1], resulting in the whole run failing (because it is intended to classify regressions/progressions, not breakages/fixes).

Even though the job is non-voting, it removed the previous Verified+1, so a committer needs to add that, or somebody needs to trigger re-check.

[0] https://logs.fd.io/production/vex-yul-rot-jenkins-1/vpp-csit-verify-perf-1908-3n-hsw/6/archives/csit_current/0/log.html.gz#s1-s1-s1-s1-s1-t1
[1] https://logs.fd.io/production/vex-yul-rot-jenkins-1/vpp-csit-verify-perf-1908-3n-hsw/6/archives/csit_parent/0/log.html.gz#s1-s1-s1-s1-s1-t1

Patch-set: 1
Reviewer: Gerrit User 1263 <1263@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>
Label: Code-Review=+1