Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> i disagree that this needs to be fixed

Are these files meant to be installed in /usr/local/etc of whatever the CMAKE_INSTALL_PREFIX is?
Even if they are only intended as templates, I believe it makes sense to have the defaults in /etc.
The user can then decide to override the defaults if desired.

Patch-set: 2