Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> (1 comment)

So initializing with &lookup, would cause longest-prefix-match to be broken.

I initialized it with NULL. This appears to be fallout from the change to use the REPLY_MACRO3_ZERO. pfx is never used uninitialized as it is only used when rv is 0, and if rv is 0 then pfx has been set.

Is there a way to make this API MP safe? I'm thinking it's not b/c of the non-atmoic lookup then fetch of the paths for the entry.

Patch-set: 5