vppapigen: api crc checker
crccheker is a tool for enforcement of the binary API.

1. Production APIs should never change.
2. An API can be can be deprecated across three release cycles.
   Release 1: API is marked as deprecated.
   option to_be_deleted_in="vyy.mm";
   option replaced_by="new_api_2";
   Release 2: both old and new API is supported
   Release 3: deprecated API is deleted.
3. APIs that are experimental / not released are not checked.

The definition of a "production API" is if the major version in the API file is > 0.

extras/scripts/crcchecker.py --check-patchset # returns -1 if backwards incompatible
extras/scripts/crcchecker.py --dump-manifest
extras/scripts/crcchecker.py --git-revision v20.01 <files>

This patch integrates the tool in "make checkstyle-api".
A future patch is required to integrate the tool in the verify job.
I.e. this patch does not enable enforcement through Jenkins.

Change-Id: I5f13c0976d8a12a58131b3e270f2dc9c00dc7d8c
Type: feature
Signed-off-by: Ole Troan <ot@cisco.com>