WIP build: enforce the .deb artifacts naming convention
This is a strawman implementation of the idea discussed on improving
the robustness of the artifact download for CSIT jobs between the releases.

They used to instll all package with the YY.MM[X]-release, just to discover that
HICN packages are tagged as YY.MM.<buildnr>-release, which caused issues.

The tweak du jour includes the "*vpp*", and something similar to this
code could be useful to keep committed to this convention from the VPP side.

Another approach could be to make "apt-get install vpp", or make some new virtual artifact that
could install all of the necessary packages..

Change-Id: Id6e5ef8c92009193dacfff88bb1898e867c778b1
Signed-off-by: Andrew Yourtchenko <ayourtch@gmail.com>