Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2: Code-Review-1
> > > 
> > > (1 comment)
> > > 
> > > Usually GCC compiler assigns enough space for enum which holds the enum values. So, I don't get the issue, you are fixing here.
> > 
> > The API generator assigns u32 to enum if not specified.
> 
> So correct fix should be to mention the correct enum size to API generator. It will be compatible with C code too.

It's really not worth it for this tiny change... Changing the enum to u8 will cause API version bump, broken CSIT tests, and it can't be patched to 20.01 where this was introduced. I'd suggest merging this patch and when there is a good reason for API version bump the vl_api_rx_mode_t can be changed to u8.

Patch-set: 2