Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> This change removes vxlan_gpe_add_del_tunnel from vpp_papi_provider.py, yet it is still used in test_vxlan_gpe.py

Yes. The method isn't removed from papi. What vpp_papi_provider does with vxlan_gpe_add_del_tunnel is wrap the papi call to provide the default values for the tests as a convenience, but this change moves the defaults to the .api so consumers beyond the tests can take advantage of them.

Patch-set: 2