virtio: put back old code in vhost_user_input to avoid regression
Some common codes were factored out to inline functions to be called by both
packed and non-packed vhost_user_input for the virtio1.1 feature. The patch
was suspected to cause vhost regression on Haswell. Let's not modifying
non-packed vhost_user_input to see if the regression is resolved.

Type: fix

Signed-off-by: Steven Luong <sluong@cisco.com>
Change-Id: If7d738a16380aa57054694130eb253d5e22d5219