Update patch set 9
Patch Set 9:

I caught a couple of behaviors that seem like not what we want:

1) ./extras/scripts/crcchecker.py --check-patchset

1a) renaming acl.api into acl_new.api passes the check. But this is *probably* an API-breaking operation ? the script does not flag it.

1b) adding a new field is correctly detected as an incompatible operation, but adding a field *and* renaming the file is not detected. 

2) ./extras/scripts/crcchecker.py --git-revision v20.01

 error: fatal: Path 'extras/deprecated/dpdk-hqos/api/dpdk.api' exists on disk, but not in 'v20.01'.

technically the condition "corresponding .api file was not on disk in the past but now there is" can be either an addition (which is fine), or a rename (which is may or may not be fine).

I think architecturally it makes sense to have a "defined_in:" as an attribute of the message, and build the list of all messages in "old" version and "new" version, with the ability to specify/restrict the source of where they come. Then the broader functions like checking the patchset and revision will be able to detect moves an changes, yet the "simple" call with just a filename will also work as it does now. Thoughts ?

Patch-set: 9
Reviewer: Gerrit User 320 <320@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>