Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Your patch is basically unconditionally disabling all advanced variants of code.
> 
> See: https://gerrit.fd.io/r/c/vpp/+/26972
> 
> init functions are defined as weak, and then I simply check if function_name != 0 before calling it...


Hi Damjan,

Thanks a lot for your reply and your feedback. I wrote this patch quite quickly to have valid gcc8 built binaries. I didn't know about weak references - nor the exact behavior of this piece of code - so this was indeed a wild guess.

If you have pointers to a way to solve this correctly that would be much appreciated.

Thanks in advance for your help

Cheers
Nathan

Patch-set: 4