Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > > Patch Set 1:
> > > > 
> > > > (1 comment)
> > > > 
> > > > Hi, Hanlin! Nice catch! However, if possible, let's try to simplify the solution a bit.
> > > 
> > > I thought about saving remaining events to unhandled_evts_vector，but because we may call some synchronous function，such as vppcom_wait_for_session_state_change， this can shuffle the order of the events.
> > 
> > That, unfortunately, holds always, not only in this corner case. Applications should not mix synchronous and asynchronous calls otherwise we can't guarantee events are delivered in order. That's why sessions that connect should be first marked as non-blocking, to avoid vppcom_wait_for_session_state_change
> 
> Not only in vppcom_session_connect, but in vcl_handle_pending_wrk_updates and vppcom_session_listen. I think the main reason for no problem may be that we will not add events to unhandled_evts_vector except SESSION_IO_EVT_RX/SESSION_IO_EVT_TX events. RX/TX events are relatively independent and will not be affected by disorder.

Normally (albeit normal here is relative to apps) vcl session worker migration should happen prior to data exchanges, i.e., I hope this is not typically done at scale for established sessions but only for listeners. So even if some events might get reordered, for individual migrating sessions it shouldn't matter much. 

Also, while waiting for the migration to complete, we keep on piling rx/tx events in the unhandled_events vector that we should eventually deliver approximately in the same order we received them. Control events like disconnects and resets will be handled out of order, but we should still try to deliver the data buffered on those sessions before notifying the app of the close event (if not, we should fix that).

Listen sessions should always be handled asynchronously at scale so those should not be a problem. With that, I think we have all the cases covered. So, for now at least, let's push everything to unhandled_evts_vector, since this a "common" solution across a multitude of vcl functions for absorbing bursts of events.

Patch-set: 1