ipsec: remove dedicated IPSec tunnels
This is a cherry-pick on https://gerrit.fd.io/r/c/vpp/+/22878
rebased to come immediately after
https://gerrit.fd.io/r/c/vpp/+/22565

Reason:
22565 started a period of time where an ipsec tests in CSIT
is failing, due to zero packets forwarded.
22878 has fixed that later, but in the meantime
the performance has changed.
In order to investigate whether the performance
is affected just by 22565 + 22878,
or whether the cause is something else merged in between,
this rebased fix has to be used.

Original commit message:

APIs for dedicated IPSec tunnels will remain in this release and are
used to programme the IPIP tunnel protect. APIs will be removed in a
future release.

see:
 https://wiki.fd.io/view/VPP/IPSec

Type: feature

Signed-off-by: Neale Ranns <nranns@cisco.com>

Change-Id: Ibe3d5b7b636f213eb7adb86b1d1de2702da706bf
Signed-off-by: Vratko Polak <vrpolak@cisco.com>