bonding: adjust link state based on active slaves
Type: improvement

When a bond interface is set to admin up, the hw interface flags are
set to VNET_HW_INTERFACE_FLAG_LINK_UP if the bond has any active slaves.
If there are no active slaves at that time and some interfaces are
enslaved and brought up later, the hw interface flags are never adjusted
when the slaves become active. So the following sequence of events
results in the bond interface being unable to send packets:

* create a bond
* set bond admin up
* set bond admin down
* set bond admin up
* add 1 or more slave interfaces to bond
* set slave interfaces admin up

When an interface is added to the list of active slaves, if it is the
first one, set VNET_HW_INTERFACE_FLAG_LINK_UP on the bond hw interface.
When an interface is removed from the list of active slaves, if it is
the last one, unset VNET_HW_INTERFACE_FLAG_LINK_UP on the bond hw
interface.

Change-Id: I31f17321f7f0e727e1ab1e01713423af6566dad9
Signed-off-by: Matthew Smith <mgsmith@netgate.com>