Update patch set 3
Patch Set 3:

What is your expectation with this feature? Do you expect it to the default for all the api's or an exceptional case?  If you expect it to become the default, is it better to invert the logic to manualendian = false, then once they are all converted, the option could be dropped.

Patch-set: 3
CC: Gerrit User 1496 <1496@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>