Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> What is your expectation with this feature? Do you expect it to the default for all the api's or an exceptional case?  If you expect it to become the default, is it better to invert the logic to manualendian = false, then once they are all converted, the option could be dropped.

The expectation is that it will be the default, but it’s the matter of getting from here to there. I suppose when all apis are marked as auto endian we could remove the flag altogether.

Patch-set: 3
Reviewer: Gerrit User 8 <8@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>