Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> >> That proves CRC checking does not take default= values into account.
> >> I think it should.
> > 
> > Why? It doesn't change the wire format, it only adds metadata to the json definition.
> 
> It changes the behavior, possibly forcing users to update their client.
> I think the worst case would be something like this (diff format):
>   - bool is_del;
>   + bool is_del[default=true];
> 
> And as CRC does not have space to distinguish backward-compatible edits from incompatible ones, it should change.
> In general, an edit to .api file can affect either papi (or other client library) users, or "direct wire" users, or both.
> Default values affect library users only. 

Library users have either the option to use the explicit value, or to use "default", which means "let the VPP decide what is best". Thus, if the VPP decided a new value is better than the old one, it is corrected and expected that no changes are needed elsewhere.

If the API owner is changing the defaults without good reason - that's a different story.
But this is where the tools won't help.

But we should probably have a good programmatic way to detect the defaults and inform the API users that they had changed, yes. But I would view that as an FYI-only activity.

> Size of enums is a complementary issue, not affecting users of higher-level client libraries.

Yes, and that's why putting multiple unrelated topics into one commit is not a good idea.

Patch-set: 7