Update patch set 3
Patch Set 3:

The code is simply incorrect, as evidenced by the JJB vote and the logs there.

If you suspect merely the prefetches are the issue, just commenting
out the body of prefetch_session_entry() in the original code should
turn it into a no-op that doesn't break anything.

Hard to say anything else given the functionality is not right.

In general - ensure you run "EXTENDED_TESTS=y TEST=acl* make test" as
a sanity check before extensive perf-tests. It's not a 100% guarantee
but it does catch a few naughty cases.

Also - take a look at f1cd92d8d9, which got about 30% improvement back
in the day, and is the source of much of the trickiness in that node.

Patch-set: 3
Reviewer: Gerrit User 320 <320@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>