Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> Hi Zhiyong Yang,
>   Why prefetch-store is done on packet data ? Whether packet data is modified in the L2 node ?
>   "CLIB_PREFETCH (p4->data, CLIB_CACHE_LINE_BYTES, STORE);" 
>   Why did you change the prefetch-store to prefetch-load ?
>   Can you please share your performance % improvement numbers ?
> 
> Thanks
> Govind

Good catch, double check that prefetch-store is a little better than load on x86 platform. Restore it in later patch. For performance, quite good performance gain can be observed on my local XEON server and I'm not sure if others can reproduce it. So the perf numbers on CSIT should be preferred and I'm trying to verify the l2 BD perf on CSIT. but it looks that CSIT system meets some trouble. It is better if anybody can help check if it has positive impact on ARM platform or not.

Patch-set: 5