interface: node recyling missing triggering graph node sync
When recycling a graph node vnet_register_interface, it is missing an
explicit call to vlib_worker_thread_node_runtime_update(). However,
there is an implicit call to vlib_worker_thread_node_runtime_update() via
vnet_sw_interface_set_flags_helper() if it enables a new feature on the
interface for the first time. But that implicit call is not guaranteed.
For example, if an interface is created, deleted, and created, then it may
skip the implicit call to vlib_worker_thread_node_runtime_update().
When that happens, the graph nodes on thread 0 are not sync'ed to the
worker threads. So the worker thread's graph nodes are out of sync
momentarily with the main thread's graph nodes until some other event
happens which calls for a sync is needed. During this window, the worker
thread's graph node is vulnerable and may experience ia crash.

Type: fix
Ticket: VPPSUPP-86

Signed-off-by: Steven Luong <sluong@cisco.com>
Change-Id: Ic9472bd2d3a212dbfeceb526506ed0400983a142