Update patch set 1
Patch Set 1:

I just think this kicks the issue down street a little bit.  Now we are taking functions that always returned instances of IP[46]Address and with this change are conditionally returning a string for an exception. This changes the exception from a ValueError to a TypeError one frame higher in the stack. The caller could just as easily catch the existing exception and do what is appropriate.  If you want to change these methods, however, wouldn't we be better to create a subclass of ipaddress._IPAddressBase?  (Which would only bump the exception higher in the stack.) Regardless, in each case, we are pushing the validation back to the caller, so why not let the caller just handle the existing exception? It is likely that each caller would have its own logic for handling the exception, no?

Patch-set: 1