Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Could you provide more details regarding what ip4 scenario when the header is not initialized?
> 
> You can try with following steps:
> 1. compile vpp with make build
> 2. create a udp session server, and listen port A
> 3. client send udp data to server port A with ip4
> 4. in the builtin_app_rx_callback, format session_dgram_hdr_t's lcl_ip or rmt_ip 
>    with format(0, "%U", format_ip46_address, &hdr->rmt_ip, IP46_TYPE_ANY)
> 5. then may get ip6 output depending on hdr0's init value.
> 
> I think it's a bug of ip_set(...), but currently init hdr0 with 0 is the fast way to fix the problem.

Actually the bug is with format_ip46_address in combination with IP46_TYPE_ANY. Don't use that because it's not accurate. Always specify the type of ip, which you actually get in the header, i.e., hdr0->is_ip4. ip_set() fills in the ip4 or ip6 part, as requested, nothing beyond that. 

So, there's no need to do the additional work of initializing the whole header to 0, because all the parts that should be initialized, are initialized by ip_set. Keep in mind that ip46_address_t is just a union of an ip4 and ip6 address, and that can't be used to accurately infer the type. 

> By the way, the hs_apps/echo_server.c not work well with udp proto and will get assert error in builtin_app_rx_callback, shoud add TRANSPORT_CFG_F_CONNECTED flags to transport_flags, like this:
> 
> u32
> udp_listen(u32 app_index)
> {
>     u32 rv;
>     vnet_listen_args_t _args = {
>         .app_index = app_index,
>         .sep_ext = {
>             .app_wrk_index = 0,
>         }
>     }, *args = &_args;
> 
>     ASSERT(vlib_get_thread_index() == 0);
> 
>     if ((rv = parse_uri(gm_ikev_main.conf.uri, &args->sep_ext))) {
>         return 0;
>     }
> 
>     args->sep_ext.transport_flags = TRANSPORT_CFG_F_CONNECTED;
> 
>     return vnet_listen(args);
> }

I assume the above snippet of code is from some code you've written? As you're pointing out, there are two modes of operation for udp, non-connected (default) and connected (must be explicitly requested). 

I believe we never changed echo server to work in non-connected mode, but it hasn't been updated recently to explicitly ask for connected service (feel free to push a patch to fix that). On the other hand, you can now use iperf3 + ldp to do udp testing.

Patch-set: 1
Reviewer: Gerrit User 193 <193@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>