misc: fix usage of lcov in extras/lcov/lcov_*
1) use lcov --zerocounters in extras/lcov/lcov_prep so running it resets
the counters (lcov --initial doesn't do this). This removes all .gcda
files recursively.

2) use the file produced by lcov --initial in extras/lcov/lcov_prep to
ensure the percentages are correct even when not all source code files
are loaded during a test. For example, the "lines touched" percentage is
the total number of lines touched during the tests, divided by the total
number of lines found in .gcda files. And .gcda files are created
whenever a source code file is reached. This means that source code
files that aren't reached during a test don't produce .gcda files, and
as a result the total number of lines was under-estimated, equivalently
the percentage of lines touched was over-estimated. Using the baseline
file (zero.info) with lcov --add-tracefile in extras/lcov/lcov_post,
this is fixed.

Tests supporting the changes:
1) --zerocounters
Pick 2 tests, e.g. ip4 and ping.
- build for gcov: (cd build-root
&& make CC=gcc CXX=g++ make PLATFORM=vpp TAG=vpp_gcov vpp-install)

- ./extras/lcov/lcov_prep
- make TEST=ip4 test-gcov
- ./extras/lcov/lcov_post
-> 14.2% lines touched

- now run another test (cumulating the coverage data):
-> lines touched: 14.9%

Now try to reset the coverage data and re-run the first test:
./extras/lcov/lcov_prep
make TEST=ip4 test-gcov
./extras/lcov/lcov_post
-> 14.9% lines touched: the coverage hasn't been reset :(

The same with the --zerocounters line in lcov_prep:
14.2% lines touched: the coverage has been reset :)

2) --initial and --add-tracefile
Most tests, maybe all, report 228978 total lines, but the baseline file
(previously named lcov.out, and unused) reports 259708 total lines
(after filtering as in lcov_post).
Assuming that this is the actual total number of lines, all global
percentages were previously slightly over-estimated.
I may be wrong here as I don't know where the missing lines are located,
maybe they're unimportant. But then the --initial line in lcov_prep
would be useless, and this is what the man page of lcov advises to do.

Type: improvement

Change-Id: Ieb475696a8d2c20002d286c5497b0ef60df80229
Signed-off-by: TimotheeChauvin <timchauv@cisco.com>