Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hi Neal, Ole,
> 
> This issue was haunting us for quite sometime in MCBU where ipv6 mpls pkt was dropping due to wrong fib index due to the above bug.
> Please help review it.
> 
> Thanks
> Rajesh Goel

Rajesh,

The VPP stable/1904 branch is unsupported.  There will be no further releases.  Any use/modification of stable/1904 should be maintained internally by your project team.  Note that stable/1908 is the current LTS branch for VPP which will be replaced by the VPP 20.09 release upon release in the September.

You appear to have submitted 3 patches for the same issue.  See [0] for instructions on how to use gerrit to modify a change.  Please abandon the duplicate gerrit changes.

This patch does not of which contain a reference to a Jira ticket, nor any indication whether the issue has been verified to be resolved on master.

I noticed that your patch failed to verify due to a make issue.  I have cherry-picked the patch [1] which resolves this issue.  You will need to rebase your patch for all CI verification jobs to pass.

Thanks,
-daw-
[0] https://fd.io/docs/vpp/master/gettingstarted/developers/gitreview.html
[1] https://gerrit.fd.io/r/c/vpp/+/27696

Patch-set: 1
CC: Gerrit User 10 <10@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>