Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > (1 comment)
> 
> Do let me know if you manage to get this tested and if all is good we should merge it.

Hi Florin,

Sorry for making you await.
My colleague run a set of loadtests and examined that the result is positive: The RX-fifo-full packet drop rate during the test (with a certain mixed-traffic profile) was 326 PPS, and it is now reduced to 101 PPS. No significant impact on throughput and CPU consumption is observed.

On the other hand, I have another not-disclosed-yet revision which avoid the "around-up" even in the case of (available_space < observed_wnd). This patch is also tested under the same loadtest. It reduced the drop rate to 0 PPS, and no significant impact on throughput and CPU consumption is observe, too. So I'd like it to be reviewed by you, but I wonder if I can share it here or I should open another review ID.

Patch-set: 3