ethernet: Support L3 dot1q subinterface without exact-match
We allow an IP address to be configured on the dot1q subinterface.
However, if exact-match option is not entered when the subinterface
is created, we don't generate the ethernet rewrite header. As a result,
locally generated packet will be transmitted without the L2 header,
causing a lot of confusions.

The fix is to remove the exact-match check and generate the proper
ethernet rewrite header even when exact-match is not configured.

Ticket: VPPSUPP-92
Type: improvement

Signed-off-by: Steven Luong <sluong@cisco.com>
Change-Id: I83035ee8f680f9379a0a8caaca2df8c1540c2c0e