Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > You don't want to change vcl layer, right?
> 
> Not necessarily, it's just that "needs real epoll" is something only needed because of ldp. Given that the vls layer is the one that handles vcl worker registrations (instead of ldp), it feels like it's not that bad that we leak epoll information into it. But, vcl should be completely oblivious, i.e., vcl used without ldp would never need this.
> 
But “needs real epoll” is introduced because of "use_mq_eventfd", and "use_mq_eventfd" is used in vcl. 😊
> 
> What I was suggesting was to completely remove need_real_epoll from vcl wrk and add it to the vls wrk. Even if one vls worker ends up registering multiple vcl workers (still not entirely convinced this is the right thing to do) this should work because only one vcl worker should be registering at a time per vls worker.

Given that one thread has registered as a vcl worker and intend to call epoll_create to create a epoll fd, and meanwhile another thread register as a vcl worker and change "user_mq_eventfd" of vls worker to  1. Then first thread will create a libc fd wrongly.

Patch-set: 2