Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2:
> > > 
> > > You don't want to change vcl layer, right?
> > 
> > Not necessarily, it's just that "needs real epoll" is something only needed because of ldp. Given that the vls layer is the one that handles vcl worker registrations (instead of ldp), it feels like it's not that bad that we leak epoll information into it. But, vcl should be completely oblivious, i.e., vcl used without ldp would never need this.
> > 
> But “needs real epoll” is introduced because of "use_mq_eventfd", and "use_mq_eventfd" is used in vcl. 😊

As I said, that should not have any effect when vcl runs without ldp. Yes, vcl asks for en epoll fd, but that syscall should just work. Given that vcl can only do that on vcl worker registration, knowing that we may steal the syscall in ldp, we must protect it externally in vls. 

> > 
> > What I was suggesting was to completely remove need_real_epoll from vcl wrk and add it to the vls wrk. Even if one vls worker ends up registering multiple vcl workers (still not entirely convinced this is the right thing to do) this should work because only one vcl worker should be registering at a time per vls worker.
> 
> Given that one thread has registered as a vcl worker and intend to call epoll_create to create a epoll fd, and meanwhile another thread register as a vcl worker and change "user_mq_eventfd" of vls worker to  1. Then first thread will create a libc fd wrongly.

That means need_real_epoll needs to be protected by a lock. When a vls worker tries to create an epoll fd, we need to guard vcl's session pool. I see we're not, but that's because for multi-process apps we have only one epoll fd per vls worker, so we never had to be careful with that. 

However, all of this begs again the question if the approach of having multiple vcl workers per vls worker is the right one. In practice, we want to treat the vls workers independently, for instance:
- the vls session pools should grow without locks
- each vls worker should have its own epoll fd (probably not more than one)
- if the vls worker has multiple threads that share the vls sessions, then multi threading locks are applied when those threads interact with the underlying vcl worker. 

Now the difficult parts:
- if vls workers want to share a session*, then somehow the session must be 1) safely cloned from one *vcl* worker to the other 2) create vls session on "new" vls worker and call vls_share_session() on the vls sessions owned by both workers. For multi-process apps both are easy because we clone both vcl and vls session pools on fork and then share the vls sessions. Doing this safely when pools can expand is hard (if they're preallocated and don't expand, things might be simpler). 
- need logic to decide when to register a new vls worker. But this you've already figured out.

So, if we can figure out how to solve the first point, which is difficult, we shouldn't need anything more in the existing infra. Any idea for your use case if this could be solved in a simple way? For instance, if the only thing shared is listener sessions, maybe those can be done upfront without much locking. If the accepted sessions are also shared, then the solution needs to be generic.  

* we can't do that with session indices, so we need something else ...

Patch-set: 2